Use ffmpeg HW accelerated video codec for stream
The video feed copy approach seems to break support for iOS 12.2 devices.
Since October 2018, ffmpeg in Raspbian benefits for hardware acceleration when using the h264_omx codec therefor making this solution acceptable for a broad audience. 
By encoding the video with it, iOS 12.2 devices should be able to once again stream.
This fixes #50.